# sdk/python/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsolidateRequest": {
        "description": "Optional knobs for the on-demand consolidate trigger.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "ConsolidateRequest",
        "type": "object"
      },
      "ConsolidateResponse": {
        "properties": {
          "dag_id": {
            "title": "Dag Id",
            "type": "string"
          },
          "dag_run_id": {
            "title": "Dag Run Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "dag_id",
          "dag_run_id",
          "state",
          "status_url"
        ],
        "title": "ConsolidateResponse",
        "type": "object"
      },
...
      "EngineVersionToggle": {
        "properties": {
          "version": {
            "enum": [
              "v1",
              "v2"
            ],
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version"
        ],
        "title": "EngineVersionToggle",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReflectionToggle": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "ReflectionToggle",
        "type": "object"
      },
...
      "SystemSettingsResponse": {
        "properties": {
          "auto_reflection": {
            "title": "Auto Reflection",
            "type": "boolean"
          },
          "engine_version": {
            "enum": [
              "v1",
              "v2"
            ],
            "title": "Engine Version",
            "type": "string"
          }
        },
        "required": [
          "auto_reflection",
          "engine_version"
        ],
        "title": "SystemSettingsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SpAIder \u2014 A multi-agent semantic knowledge graph platform powered by LLMs and Neo4j. Ingest unstructured text, build a persistent knowledge graph, query it in natural language, and synthesise fine-tuning datasets for downstream models.",
    "title": "SpAIder",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/system/consolidate": {
      "post": {
        "description": "Trigger the graph_maintenance Airflow DAG immediately, off its\nnormal schedule. Useful right after a heavy ingest, or when the\noperator wants to run consolidation without waiting for Sunday 03:00 UTC.\n\nBehaviour\n---------\n- Returns the new ``dag_run_id`` and a ``status_url`` the operator can\n  poll (or open in the Airflow UI).\n- Returns **503 Service Unavailable** when ``AIRFLOW_BASE_URL`` is\n  unset or unreachable. The error body points at the CLI fallback\n  (``python -m app.scripts.run_consolidation``) which works without\n  Airflow.\n\nAuthentication: re-uses ``AIRFLOW_USERNAME``/``AIRFLOW_PASSWORD`` from\nsettings (basic auth \u2014 Airflow's stable v2 default). Airflow's other\nauth schemes (Kerberos, JWT) would need a separate adapter.",
        "operationId": "trigger_consolidate_api_v1_system_consolidate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ConsolidateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsolidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Consolidate",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/v1/system/settings/engine": {
      "post": {
        "description": "Switch the active graph engine between V1 and V2.\n\n- **v1**: Classic RAG graph \u2014 all edges rendered with uniform width.\n- **v2**: Cognitive Graph \u2014 edge width reflects `utility_weight` synapse strength.\n\nThe setting is persisted in Neo4j and immediately visible to all connected clients\nvia the EngineContext (frontend polls or reads on mount).\nReturns 404 if the settings node is missing (call GET first to initialise).",
        "operationId": "set_engine_version_api_v1_system_settings_engine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngineVersionToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Engine Version",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/system/settings/reflection": {
      "post": {
        "description": "Enable or disable the autonomous Reflection Engine (Hippocampus).\n\nThe toggle is persisted on the global SystemSettings singleton in Neo4j\nand is checked by the background scheduler every 5 minutes.\nReturns 404 if the settings node is missing (call GET first to initialise).",
        "operationId": "set_reflection_toggle_api_v1_system_settings_reflection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReflectionToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Reflection Toggle",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}